requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8,<3.0
djangorestframework>=3.9,<4.0
djangorestframework-simplejwt==4.4.0
drf_yasg==1.17.1